Files to fit curve
This calls python scipy/curvefit to fit an exponential curve (Function: Y = A0 + (A1 * exp(-1/t1)) + (A2 * exp(-t/t2))). 
We set the parameters so A0 + A1 + A2 = 1
The do_curve_equ.sh tries different configurations.
get_best_curve_equ.sh finds the best curve based on the lowest sum of the squares of the differences between the curve created and the data. It creates a function file with the values fo Ax and tx and a plot file for the fitted curve.